requirements.txt
numpy>=1.15.1
scikit-learn>=0.19.2
python-mnist>=0.6
pyyaml>=3.13
scipy>=1.1.0
matplotlib>=2.2.3